requirements.txt
numpy
h5py
unyt
velociraptor-python
scipy
matplotlib
sphviewer